change calc_average function return type and update related calls

+++ cpu/inc/utils.h
@@ -36,7 +36,7 @@ static volatile uint32_t _high, _low;
 
 
 void measure_cpufreq();
-uint64_t calc_average(uint64_t* ticks, int iterations, int min, int max, int diff);
+double calc_average(uint64_t* ticks, int iterations, int min, int max, int diff);
 
 double_t calc_timeread_overhead(uint64_t iterations);
 double_t calc_loop_overhead(uint64_t iterations);
